pyproject.toml
[tool.poetry]
name = "credit-card-airflow-example"
version = "0.1.0"
description = "Airflow project that uses credit card data to demonstrate basic ETL concepts"
authors = ["Collin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
apache-airflow = "^2.10.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
pyarrow = "^17.0.0"
tortoise-orm = "^0.21.6"
psycopg = "^3.2.2"
asyncio = "^3.4.3"
asyncpg = "^0.29.0"
sqlalchemy = ">=1.4.36,<2.0"
psycopg2 = "^2.9.9"
boto3 = "^1.35.28"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
